+++ utils/printUtils.hpp
@@ -11,6 +11,6 @@ char* getString(char* buf, int max);
 
 int stringToInt(const char* string);
 
-void printInt(int xx, int base = 10, int sgn = 0);
+void printInt(uint64 xx, int base = 10, int sgn = 0);
 
 #endif //PRINTUTILS_HPP
